edit domain tests
- rename BuildingDomain to BuildingNonBusinessTypes
- remove DomainTestSpec
- move setup and helpers to BuildingDomain test fixture
